Rename Treasure_t type to GameObject_t
I made this change because Treasure does not seem a good name for this
type. The object data list in treasure.cpp not only contains treasure
items, but also things like stairs and rubble. It can also make it
confusing when looking at the array containing treasure items
(inventory) found on the level.

This commit also renames the global object_list[] array to
game_objects[].